+++ src/main/java/net/oneandone/stool/classes/Expressions.java
@@ -22,7 +22,6 @@ import freemarker.template.TemplateModelException;
 import freemarker.template.TemplateSequenceModel;
 import net.oneandone.inline.ArgumentException;
 import net.oneandone.stool.core.LocalSettings;
-import net.oneandone.stool.util.Expire;
 import net.oneandone.sushi.fs.World;
 import net.oneandone.sushi.fs.file.FileNode;
 import net.oneandone.sushi.launcher.Launcher;
@@ -164,8 +163,6 @@ public class Expressions {
         result.put("fqdn", fqdn);
         result.put("stage", stage);
         result.put("host", host);
-        result.put("defaultExpire", Expire.fromNumber(localSettings.defaultExpire).toString());
-        result.put("defaultContact", "");
         result.put("workdir", (TemplateMethodModelEx) list -> {
             if (list.size() != 1) {
                 throw new ArgumentException(list.toString());